net - use ffcanvas temp files instead of Document's complicated solution

+++ src-net/TempFileDeleter.cpp
@@ -25,32 +25,31 @@
 
 #include "StdAfx.h"
 #include "TempFileDeleter.h"
-#include "tempfile.h"
 
-TempFileDeleter::TempFileDeleter(tempfile_ptr file, System::Diagnostics::Process^ viewer)
-    : mTempFile(nullptr), mViewer(viewer)
+TempFileDeleter::TempFileDeleter(System::String^ file, System::Diagnostics::Process^ viewer)
+    : mTempFile(file), mViewer(viewer)
 {
-    if (viewer && !viewer->HasExited && file && file->written()) {
-        mTempFile = gcnew System::String(file->name().c_str());
-        file->release();    // this object owns the file now
+    if (viewer && !viewer->HasExited) {
         TempFiles->Add(this);
         viewer->Exited += gcnew System::EventHandler(this, &TempFileDeleter::DeleteIt);
-    }
-    if (viewer && !viewer->HasExited)
         viewer->CloseMainWindow();
-    // tempfile_ptr file goes out of scope and is deleted, temp file is deleted
-    // too unless this object takes over ownership of the file
+    } else {
+        DeleteIt(this, nullptr);
+    }
 }
 
 void
 TempFileDeleter::DeleteIt(Object^ sender, System::EventArgs^ e)
 {
-    try {
-        System::IO::File::Delete(mTempFile);
-        System::Diagnostics::Debugger::Log(3, "Delete success: ", mTempFile);
-    } catch (System::Exception^) {
-        System::Diagnostics::Debugger::Log(1, "Delete failed: ", mTempFile);
+    if (mTempFile) {
+        try {
+            System::IO::File::Delete(mTempFile);
+            System::Diagnostics::Debugger::Log(3, "Delete success: ", mTempFile);
+        } catch (System::Exception^) {
+            System::Diagnostics::Debugger::Log(1, "Delete failed: ", mTempFile);
+        }
     }
     mViewer = nullptr;
-    TempFiles->Remove(this);
+    mTempFile = nullptr;
+    TempFiles->Remove(this);    // Does nothing if never added to List
 }
